backend/app/agents/requirements_openai.txt
# Requirements for Playwright + OpenAI + LangGraph (model_dump compatible)
# This combination has been tested to work without model_dump serialization errors

# Core LangGraph and LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain>=0.3.0

# OpenAI integration
langchain-openai>=0.2.0
openai>=1.0.0

# Playwright for browser automation
playwright>=1.40.0

# Pydantic (should work with v2.x)
pydantic>=2.0.0

# Standard dependencies
python-dotenv>=1.0.0
asyncio-compat

# Optional: For better debugging
rich>=13.0.0